works

+++ system/bin/generate_html.sh
@@ -1,7 +1,11 @@
 #!/bin/bash
 
-echo "Creating symbolic link /workspace to $(pwd)"
-ln -s "$(pwd)" /workspace
+# Create a symbolic link to the current directory
+# if it doesn't exist
+if [[ ! -L "/workspace" ]]; then
+    echo "Creating symbolic link /workspace to $(pwd)"
+    ln -s "$(pwd)" /workspace
+fi
 
 # Check if php-fpm is running
 if ! pgrep "php-fpm" > /dev/null; then
@@ -21,9 +25,10 @@ fi
 
 echo 127.0.0.1 "$hostname" >> /etc/hosts
 
+# Curl the URL $1 and save it to $2
 function curl_and_save() {
-  echo Hitting URL "$1"
-  curl -k "$1" > "$2"
+  echo "curl_and_save https://$hostname$1 > $2"
+  curl -ks --fail "https://$hostname$1" > "$2"
 
   # Check if the file exists and is not zero bytes
   if [[ ! -s $2 ]]; then
@@ -31,14 +36,18 @@ function curl_and_save() {
       exit 1
   fi
 
-  # Check if the file is missing the specific text string
-  if ! grep -q "<!-- Built with Ata's SSG https://www.github.com/atas/ssg -->" "$2"; then
-      echo "Error: The file does not contain the attribution HTML comment OR something else is wrong in the output. As the MIT license requirement, please do not remove it to help other fellow developers discover this neat tool."
+  # Check if file exists
+  if [[ ! -e "$2" ]]; then
+      echo "Error: File $2 does not exist."
+      exit 1
+  fi
+
+  # Check if file size is zero
+  if [[ ! -s "$2" ]]; then
+      echo "Error: File $2 is empty."
       exit 1
   fi
 }
-
-curl_and_save "https://$hostname/" "html/index.html"
 
 #region Building HTML files
 rm -rf html
@@ -47,7 +56,7 @@ mkdir -p html
 # pages/*.md HTML BUilding
 for file in pages/*.md; do
   filename=$(basename "$file" .md)
-  curl_and_save "https://$hostname/$filename" "html/$filename.html"
+  curl_and_save "/$filename" "html/$filename.html"
 done
 
 # posts/*.md HTML Building
@@ -55,12 +64,13 @@ mkdir -p html/p
 for file in posts/*.md; do
   filename=$(basename "$file" .md)
   slug=$(awk -F': ' '/^slug:|^[ \t]*"slug":/ {gsub(/["\r]/, "", $2); print $2}' $file)
-  curl_and_save "https://$hostname/p/$slug" "html/p/$slug.html"
+  curl_and_save "/p/$slug" "html/p/$slug.html"
 done
 
 # **/*.php Any php file HTML Building
-find . -name "*.php" ! -path "./system/*" | while read -r file; do
-    if [[ "$file" == "index.php" || "$file" == "page.php" || "$file" != "post.php" ]]; then
+find . -name "*.php" ! -path "./system/*" ! -path "./vendor/*" ! -path "./layout/*" | while read -r file; do
+    file="${file#./}" #Remove the leading ./
+    if [[ "$file" == "page.php" || "$file" == "post.php" ]]; then #pages and posts are handled above separately
       continue;
     fi
 
@@ -72,7 +82,7 @@ find . -name "*.php" ! -path "./system/*" | while read -r file; do
     mkdir -p "$(dirname "$html_file")"
 
     filename=$(basename "$file" .php)
-    curl_and_save "https://$hostname/$without_extension" "$html_file"
+    curl_and_save "/$without_extension" "$html_file"
 done
 
 #endregion
